__CPROVER_llabs returns a long long
Fixes the declaration in builtin_headers.h. Type checking actually
ignored this declaration and would always generate the expression of the
correct type.

+++ src/ansi-c/cprover_builtin_headers.h
@@ -84,7 +84,7 @@ int __CPROVER_isunorderedd(double f, double g);
 // absolute value
 int __CPROVER_abs(int x);
 long int __CPROVER_labs(long int x);
-long int __CPROVER_llabs(long long int x);
+long long int __CPROVER_llabs(long long int x);
 double __CPROVER_fabs(double x);
 long double __CPROVER_fabsl(long double x);
 float __CPROVER_fabsf(float x);
